kdevops: add q35 support
libvirt uses the 'pc' type by default to this day. WTF right?
Well we can't use 'pc' as that uses i440x and that's ancient
without PCIe. So we can't do pcie-hotplug or CXL. So add support
for that.

All we gotta do is just ensure we have pcie-bridge added which
we also use for the nvme drives just as we do for the old school
i440x stuff.

Long ago I had made this a default but that turns out to be a bad
idea because some images have old defaults set, so users should
be explicit about this need, or have kconfig options select or
depend on q35.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>