xfs: add generic/388 as failing for 5.10.y config xfs_reflink_1024
The test is failing almost every run on config xfs_reflink_1024
on kernel v5.10.131.

From longer tests with previous 5.10.y kernels, the test was reported
to fail at lower rate for other config, but leave it in the specific
config section for now to see if failure rate has changed with other
configs.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>