vagrant: use string instead of int for zone capacity kconfig
string is used for zone size but int is used for zone capacity in the
kconfig option.

Using a int value for zone capacity such as 3 will be interpretted as 3
bytes instead of 3M, which is typically what we want.

Use string as Kconfig input instead of int with "0M" as the default
value for zone capacity.