gen_nodes: ensure proper permissions on nodes and vagrant file
When the gen_nodes ansible role was used sudo was used to generate
the nodes file and Vagrant file by mistake, that's not needed.
So ensure we revert to the regular user permissions. We can perhaps
later remove this check / setting later, way later after more users
deploy this.

We need this as otherwise subsequent management of the files as a
regular user would not work, which we'll do next for other reasons.

This should introduce no functional changes other than ensuring
the following files have the same permissions as the regular user:

  * vagrant/kdevops_nodes.yaml
  * vagrant/Vagrantfile

Fixes: f81a6593ff2b ("kdevops: generate nodes & Vagrantfile file with jinja2 templating")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>