gen_nodes: provide a guard to protect Vagrantfile
The way vagrant works is it uses the Vagrantfile used to
instiate nodes and puts it into a libvirt storage pool,
and if the file changes it'll barf treating the new file
as a new instatiation effort. That means new features cannot
be appended as kdevops advances.

So ensure to keep the Vagrantfile static always and never
replace it. If a user wants to override this, they can rm -f it
themselves, no need to add extra complexity like a `make FORCE=1`
or anything like that, when we can simply ensure the users
understand this limitation and its respective implication by
requiring removal of the file manually.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>