Workflows: Add a new test message id workflow
When encountering patches in kernel mailing lists it is advantageus to
check them against a predefined baseline to ascertain if there are any
regressions. This check based on message id is enabled by this commit.

When enabled, the b4 command from public-inbox will rebase the patches
contained in the message-id on top of a baseline kernel and install it in
the development box. A baseline kernel without the patches will be
installed in the baseline box. The effects of the patch on the tests can
then be seen as the difference between the baseline and development test
results.

To enable this workflow the "Enable both a baseline and development systems
per target tests" as well as "Test mailing list message ID" Kconfig options
need to be enabled.  Additionally the message id must be placed in the "The
message ID" option.

Signed-off-by: Joel Granados <j.granados@samsung.com>