bootlinux: add an install b4 path for rh/suse
Commit 6a38ae72e349b ("Workflows: Add a new test message id workflow")
added support to use b4 to apply a pending patch. This also added
support to install b4, but it did so only for debian. The way I go
about thing is I always strive to add support for at least:

  * debian (and therefore ubuntu)
  * rh (and therefore fedora)
  * suse (and therefore opensuse)

Since the b4 install commands are the same for all other distros
just generalize that install path as shared.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>