kdevops: add custom storage pool name support
By default vagrant uses and relies on the fact that libvirt would
have created a default pool named "default". Creating custom paths
for nvme drives uses a custom set of commands so we don't use libvirt
to create these drives from this pool path, however we still use the same
path unless you modify the default CONFIG_LIBVIRT_STORAGE_POOL_PATH_CUSTOM.

So if you modify CONFIG_LIBVIRT_STORAGE_POOL_PATH_CUSTOM you get your
nvme drives on that path with libvirt and also the CXL stuff, but not
the main guest image.

To fix this we must play with libvirt directly and create a custom pool
if the user wants one. Add an ansible role to deal with this for users
so they don't have to do the work themselves.

With this added you just need to enable now LIBVIRT_STORAGE_POOL_CREATE
and set the LIBVIRT_STORAGE_POOL_NAME to something other than "default"
to take advantage of this. This ensures you use all of the guest data
on a custom path set by LIBVIRT_STORAGE_POOL_PATH_CUSTOM.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>