kdevops: add support for raw drive format type for libvirt
Amir Goldstein is reporting that nvme timeoutes are happening
on XFS when testing with fstests on the punch hole group when the
backing drive for the guests is over 80% filled. That is, not the
guest drives, but the real nvme drives where the guest nvme sparse
qcow2 files live.

We had heard before of issues with qcow2 and zone resets
triggering nvme timeouts follows even by RCU CPU stalls [0]
and during that discussion it was clear that there may be
qemu bugs which make its implementation of zone resets not as
optimimal. The way zone resets work on qemu is it uses write
zeroes (REQ_OP_WRITE_ZEROES) so to ensure blocks in an empty
zone must be considered deallocated. It was believed that this
approach can create delays as it means actually writing the
data on the drive, and Shinichiro suggested for qemu instead to
use fallocate with ALLOC_FL_PUNCH_HOLE | FALLOC_FL_KEEP_SIZE.
At least as per my review falloca with ALLOC_FL_PUNCH_HOLE |
FALLOC_FL_KEEP_SIZE does end up in Linux blkdev_issue_zeroout()
with BLKDEV_ZERO_NOFALLBACK. As per my review this also just
ends up calling __blkdev_issue_write_zeroes() with
REQ_OP_WRITE_ZEROES. So it is not clear what the recommendation
was then.

Since Amir is running into issues with nvme timeoutes also on
fstests punch hole tests I'm inclined to believe the reason to
these issues issue is actually different. qcow2 uses a cluster
of 64 KiB size, and by as per qemu commit 54064e51d1e2
("hw/block/nvme: add dulbe support") "by default and blocks will
only be marked deallocated if a full cluster is zeroed or discarded".
This means a punch hole on a filesystem for 4 KiB may end up in
non deterministic delays until the full 64 KiB cluster size
is also deallocated. My hypothesis then is if a guest file is on
a filesystem which is close to being filled (say over 80%) full
the backing storage drive may end up doing nonsense while it
can ensure to deallocate the entire 64 KiB block size on behalf
of the guest on the host filesystem for the guest file. For this
reason I'm inclined to even suggest we verify qcow2 format files
should not inform the upper layers to support hardware offload
for zeroing. It is not clear if this is the assumption today.

Either way, this needs more testing, and this enables the testing
of raw devices with libvirt on kdevops.

[0] https://lore.kernel.org/all/YmkFAYXL/xgBSMA8@apples/

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>