fstests: support setting fstests_section to "all"
For specialized vms we may want to give it host specific configs and
then run all of the sections, not just the host inferred section.  Add
the logic to handle setting fstests_section to "all", which will simply
run ./check with the appropriate config, and then check will run all the
sections in that config.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>