fstests: do not re-init extra_args.yaml when doing make fstests-baseline
This dependency was added to all of the fstests make targets, but it
appears we only need it for the CI loop target.  The idea was that if
we're looping through multiple FSTYP's we need to re-init the local
arguments to their appropriate values.  However having this target set
means that any user provided overrides are overwritten.

For now just apply this to fstests-baseline, but we may want to remove
this dependency from all of the make targets except for
fstests-baseline-ci.  I don't know why this was originally added so I'm
making this change as narrow as possible.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>