fstests: copy udev rule to change sched for btrfs zoned tests
For some reason, the udev rule that gets packaged with btrfs-progs[1]
does not update the scheduler correctly when the block device is
formatted with btrfs. However, the existing udev rule for changing
scheduler works if the zoned device is already formatted with btrfs
while booting.

To avoid this issue, copy an adjusted udev rule which always changes the
scheduler for a zoned block device to mq-deadline. This is done only for
btrfs fstetsts for zoned devices as some filesystems may not require this.

[1]https://github.com/kdave/btrfs-progs/blob/master/64-btrfs-zoned.rules
Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>