bootlinux: bump linux-next to next-20220930
The only thing I see worth enabling is CONFIG_LRU_GEN and its
respective kconfig which enables it by default CONFIG_LRU_GEN_ENABLED.
As described in the helper for CONFIG_LRU_GEN_STATS the stats should
only be enabled if debugging and we're not doing that for this stuff
so disable the stats.

Rationale for enabling this is the description of the amount of
workloads and gains / third party evaluation is impressive, and
a few distributions *already* had merged this and had been testing
it for a while. Time to get up with times and test this too by
default.

I'd like to see what rationale folks come up with to disable this
for all the tests we support.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>