bootlinux: bump to from v6.0-rc7 to v6.0-rc7
Add v6.0-rc7 and make it the default for Linus' tree.
Instead of a symlink add a full file this time only to
demo the difference with what was done for v6.0-rc6 which
Amir updated using a symlink to v6.0-rc1, but there is really
no need to do this full file update as there are no changes.
The only change with using this mechanism is that if you are
manually doing 'make oldconfig' you can more easily *only* see
the things which did change in between v6.0-rc7 and v6.0-rc8
once you run 'make oldconfig' on v6.0-rc8 or if that RC release
is skipped on first real v6.0 release.

Using symlinks means you will have to review each time all
the changes since the first symlink. The only thing merged
worth mentioning which are are disabling on purpose here is
CONFIG_LRU_GEN and CONFIG_LRU_GEN_ENABLED. But I've enabled
that on linux-next kernels configs to start to help test that
as it would seem that is the 'way of the future'.

Let's learn with what Amir did on kdevops commit f4cd4423a69f
("bootlinux: bump the default V60 kernel to v6.0-rc6") with
respect to the expunges where he made a symlink the v6.0-rc1
expunges for fstests to rc6. Since we are not making any
significant changes to the kernel config we can be sure
the same expunges can be used for rc7 as with rc6.

Let's also augment this practice then to also augment the
expunges for blktests from 6.0-rc1 to symlink to 6.0-rc7.
That should enable folks testing 6.0-rc7 to also gain the
benefits of using the same expunges as 6.0-rc1.

Moving forward hopefully folks will pick up on this and
do this as kernels get updated. The rule of thumb should be
that if we are not doing huge kernel config updates that
would affect a baseline we should be able to keep the same
expunges.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>