git-daemon@.service: fixes to get this to work
Socket activation works by only starting a service if we get tickle on
a socket. This allows us to save resources and stop having to start
stupid services without them being around. So from a security
perspective we need to use DynamicUser=yes now so that a stupid
ephemeral user is used for this, and --inetd so we don't have to
type in the exact IP addresses / hostnames we want to host this
for.

I've tested this against network hosts on the same network and also
by guests on the hypervisor hosting the git mirror. It works as
expected.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>