blktests/gen-results-dir.py: replace distutils.dir_util as per PEP632
As per PEP 632, we end up with a deprecation warning now for usings
distuils, in particular for the documentation for PEP 632 it tells us
we must provide our own re-implemtation for using distutils.dir_util and
so we do this with shutils.copytree(a, b, dirs_exist_ok=True) and a
check for the source directory to exist before.

This fixes the error on:

/data2/blktests/kdevops/playbooks/./python/workflows/blktests/gen-results-dir.py:13:
DeprecationWarning: The distutils package is deprecated and slated for
removal in Python 3.12. Use setuptools or check PEP 632 for potential
alternatives

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>