blktests: add sanity check for blktests_test_devs
When blktests_test_devs is not properly set ./oscheck.sh will fail
as it is a setup issue but since we ignore errors we don't get told
what happened. So we must do some part of the sanity check's that
blktests check does itself.

Do that and be pedantic so it is clear what is going on.

I discovered then when testing a q35 guest which had /dev/nvme2n1
defined as the blktests_test_devs but yet we have not created that
device on the guest. When that happens ./oscheck.sh would fail but
since stderr is ignored we don't set the setup error as we ignore
all test failures.

Fix this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>