vagrant: switch to Fedora 37
...and simplify Kconfig.fedora.

VAGRANT_FEDORA_35_35202110260 is not used outside of the Kconfig.fedora.
It's only used to select a single version. All of the latest Fedora
versions should support all of the xfs features. Squash this down into a
single stanza for now.

Is there a universal "latest cloud image" name we could use instead?
It'd be nice if this updated on its own.

Signed-off-by: Jeff Layton <jlayton@kernel.org>