fstests: fix enabling custom tests
It turns out that upstream fstests check.sh *will not limit the scope
of your tests if you are using a group, and so if you enable custom
tests today, kconfig won't disable the default auto group crap. So
you end up scratching our head why variables are not being respected.

This is a bug introduced by da523b056475 ("fstests: add config for
running a custom set of test") in its use of the oscheck_extra_args
under the assumption users will *read* the kconfig stuff.

The way to properly set defaults first, is by using the ansible roles
default/main.yml file. Then, to ensure folks don't do something stupid
a kconfig bool is needed and the crap you don't want to enable must then
be negated by that bool.

The auto group crap is shoved down our throughts today on oscheck.sh
and so we need to pass along a new environment variable which if set
will completley disable the group crap.

We use a sensible variable which we can later expand for dynamic support
for limitting test scope.

Fixes: da523b056475 ("fstests: add config for running a custom set of test")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>