fstests: add support to dynamically limit tests
Often times you may want to just limit the amount of tests you run,
but not modify your test configuration with kdevops. To do this you
can use the environment variable LIMIT_TESTS. For example:

make fstests-baseline HOSTS="next-xfs-crc" TESTS="generic/002 generic/003"

This adds support for this.

The next feature we can work on is to read the expunges for reach
host and use that for reach target run on each host to verify that
the expunges are still valid. We then look for the inverse, we
are looking to see what tests are now passing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>