Fix multiple limitted hosts
Using HOSTS to limit your host scope needs to be done in a way
that is portable and right now I'm seeing that '-l foo -l bar" will
just run the playbook on the last argument on the list, so we need
to just use '-l foo,bar'. So embrace that.

Fixes: 271cbd7a ("Makefile: add support to limit hosts")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>