oscheck-lib.sh: add helper to set the RUN_SECTION
The RUN_SECTION tells us what section we want to use from
the fstests's configuration file. We typically process this
through the script's own argument processsor, so have a helper
which let's us give it what the user passed so other scripts
can also make use of this.

If nothing is passed we infer a section based on the hostname.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>