oscheck-lib.sh: move ONLY_QUESTION_DISTRO_KERNEL to lib init
This variable is used to verify if we are only running a distro
kernel or not. It is usefulf for enterprise kernels which want
to perhaps only allow tests to run with distro specific kernels
and not custom crap.

Move initialization of this variable to the library init for oscheck.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>