fstests: fix usage of the XFS realtime device
The realtime device configuration was missing  set of things.
Fix this, and let the blowing up beging. Who's using the realtime
device? Great question, however it would seem that it's not stable
and so let's use public resources to infer the proper way to set up
the realtime device with -r rtdev=foo -d rtinherit=1 as suggested by
Dave Chinner back in 2017 [0]. This invalidates our rtdev baselines
and so we must start fresh too there. But then again I haven't been
testing rtdev and I am not aware of anyone on the planet properly
testing these things either.

While at it, remove the stale [xfs_realtimedev] profile, as that never
matches any of the hostnames we define so just kill it. The real profile
is just [xfs_rtdev].

The new rtdev uses /dev/loop13 as it is free.

Drop the external journal for it as well. If we want to test that we
should do that separately.

We drop LOGWRITES_DEV usage too as that seems to create a conflict,
but note that has nothing to do with an external journal. I'll update
documentation about that shortly.

[0] https://patchwork.kernel.org/project/xfs/patch/25856B28-A65C-4C5B-890D-159F8822393D@fb.com/
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>