xfs: add support for new target: rtdev+logdev = xfs_rtlogdev
Why would you use both the realtime device *and* an external journal?
Some folks seem to be using the realtime device as a cheap way to do
some form of data tiering, even though the consensus from the community
is that this is a really inproper way to do this [0]. The recommended
mechanism that folks recommend to do that is to do futur enhancements
to a specialized allocation group in XFS and for you band together
drives using device mapper.

Why would you use an external journal on top of that? Well it is known
that for fsync intense workloads it can help to about 50% of IOPS so
if you want to do both the above and get the benefit of an external
journal you'd use this crazy setup.

Let's provide a test target for this and see what sort of crap blows up.

We need to provide a new scratch dev pool which skips /dev/loop12 so
we can re-use that for all the TEST_RTDEV.

[0] https://patchwork.kernel.org/project/xfs/patch/25856B28-A65C-4C5B-890D-159F8822393D@fb.com/
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>