vagrant: enable libvirt CPU 'host-passthrough'
As per qemu's documentation the preferred CPU model is
to be set to host-passthrough, which just passes along
the host CPU model features, model, stepping, exactly to
the guest. This is the recommended CPU to use, provided
live migration is not required [0], and since kdevops
doesn't need live-migration yet let's make this the default
moving forward.

[0] https://qemu.readthedocs.io/en/latest/system/qemu-cpu-models.html

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>