bootlinux: add a more sensible default host path for 9p
If you git clone Linux using 9p you're doing that on the host,
right now we use the home directory but in consideration of
having different instances of kdevops run on the same account
it would be silly to use the $HOME path as you may wan to
use the same account to try to build different versions of
Linux.

Use the PWD/linux/ target for the build, that way we already
use the local kdevops namespace scheme and use the IO on the
same drive.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>