vagrant / bootlinux: clone linux prior to bringup for 9p support
When doing 'make bringup' with 9P support enabled, in order to bring up
the guests we need the linux clone already completed so we can share it
with the guest.

So provide support to clone linux *first* automatically for you
on the host if you've enabled 9P support.

While at it, this provides the generic 'make linux-clone' target
which can be used whether or not you did enable 9P support. If
you disabled it, the clone will happen on the target guests,
otherwise it'll happen on the host. With a tiny bit of work
I'm sure we could enable an option to *force* cloning on the
guests even if we enabled 9P just for kicks, in case we needed
or wanted that, but this is enough of a big change to support
for now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>