bootlinux: ignore LIMIT_TESTS when issued against host
The 'make linux-clone' will resolve to the linux-clone-9p target
when we want to git clone linux and 9P is enabled. When that is
done the localhost is used, so we must drop the LIMIT_HOSTS
otherwise we end up trying to fetch information from the clients.
The clients may not be ready yet as we now issue the git clone
*prior* to bringup to support the 9p mount tag. The git tree
for Linux must exist prior to instantiating the clients.

Fix this by dropping LIMIT_HOSTS for the linux-clone-9p target.

Fixes: 5c3edcddd1e4 ("vagrant / bootlinux: clone linux prior to bringup for 9p support")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>