fstests: enhance labels used when looking for failed tests
When we look for failed tests files we provide a short name,
but this isn't that helpful when debugging so just provide the
full exact name used. We fix the label used for dmesg files
tests_dmesg_name as that is actually what is being used.

Note: If a test fails but not due to a real failure, no dmesg
file is generated :) but at least now I feel pretty confident
the file *is* properly looked for.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>