fstests: fix parsing of xfs_logdev and xfs_rtlogdev mkfs params
The mkfs params for the external logdev have a -lsize=1g setting,
but our processing of yaml entries looks for all '=' and replaces
them with ':' and this breaks its use with mkfs. To fix this
add a new extra vars processing entry which let's us do the regexp
of the s|=|:| only once to the first entry. A little bit of Make
vodoo is used.

We may be able to just use this later for all processing but best
actually would be to get upstream into kconfig yaml output support.

Fixes: 0a1e8d05 ("fstests: enhance external log testing")
Fixes: acf483c2 ("xfs: add support for new target: rtdev+logdev = xfs_rtlogdev"
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>