btrfs: expand profiles for btrfs
Josef provided a set of btrfs profiles long ago when I asked how we
should test btrfs during some LSFMM thread. The set of profiles Josef
had different names, and were of a smaller set but in it was enough
examples to enable us to have grow this systematically.

The main features are, no-holes, and the free space tree. Then the
different set of compression algorithms: zlib, lzo, zstd. To test
old kernels we can use a simple term "holes", which essentially is
just an acknowledgement that it lackes "no-holes" and also lacks the
free space tree.

There's the one-off "dave" profile Josef carried, that essentially is
the btrfs_fspace_dsync for synchronous discard with the free space
tree and no-holes disabled. I'll keep that only for the fspace tree
but it's easy to grow this to the other main profile sets if this
turns out to be really useful.

RAID profiles are not covered yet, because well, no enterprise Linux
distribution supports that yet. However, the work put into this commit
should help pave the way to easily expand support to add support to hep
test for any btrfs feature.

The *one* perf test requires setting PERF_CONFIGNAME and so we now set
that. We use the kdevops version for that, but also the target kernel
version if its an upstream built kernel, otherwise we just peg "distro"
to it.

To see what this ends up producing without instantiating clients or
guests you can use:

  make fstests-config-debug

That should generate the file that would be used on the clients / guests on
workflows/fstests/btrfs/btrfs.config.generated

And now we have 19 btrfs target test profiles.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>