fstests: provide a namespace for results
We currently put results into workflows/fstests/results/*.xz
however this doesn't make much sense if we want to scale.
Provide a namespace for contributors and also for *how* the
clients or guests are instantiated. For instance, we can use
libvirt or a cloud provider.

This also let's use have different contributors help tests the
same kernel.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>