libvirt: add initial dynamic PCI-E passthrough support through kconfig
Kconfig typically is used in a static form, however, it isn't hard
to generate dynamic Kconfig files either, we just need to hook right
into before the target config interface is compiled to generate our
own files. We need this to support a nice way to allow users to select
which PCI-E device they want to use for PCI-E passthrough support onto
libvirt guests.

This adds initial support for this. "Initial support" means that the
grunt of the work is done now, and all we need now is to finalize a
PCIE topology we are happy with, and test this feature. I haven't gotten
it to work yet, but at least the output of the nodes file matches the
documented expectations.

To use it use:

  make dynconfig

This is just a short cut for:

  make menuconfig KDEVOPS_ENABLE_PCIE_KCONFIG=1

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>