kconfig.Makefile: allow dynamic Kconfig generation
Although it is not popular yet, dynamic kconfigs are indeed possible
and quite useful. An example useful is to support scraping PCI tables
to allow PCI-E passthrough management to virtualization guests using
a framework. That use case is being developed in kdevops now so enable
it's implementation by allowing the user of this git tree to adding
new dependencies onto the top level Kconfig being read. We add that
dependency to the Kconfig file itself and so easily allowing such
dependency to obviously be nothing as the file already exists in
those projects already using kconfig as-is. Projects which want to
use dynamic kconfig generation can either generate their Kconfig file
always now, or simply add dependencies to their Kconfig file with
subsequent child dependencies.

The more complex trick is *how* to use these generated files at will,
for that we can leverage the trick implemented in kdevops for allowing
git trees to take kdevops as a gitsubtree, expand on it and expand their
own Kconfig symbols. This was done to support SUSE Enterprise Linux
testing with kdevops, which used kdevops as a git subtree, and so the
sle-kdevops tree would expand on vagrant/Kconfig.suse which would reside
as an empty file on kdevops. Then you add HAVE_* config variables which
check to see if the file is empty or not to enable that kconfig symbol
as a bool to "y" or "n"

config HAVE_FEATURE_BAR
	bool
	default $(shell, scripts/check_kconfig.sh foo)

The script would check with -s for the Kconfig.foo and
echo "y" or "n"

You can then use something like this to dynamically and optionally
input your new generated Kconfig file:

if HAVE_FEATURE_FOO
source "Kconfig.foo"
endif # HAVE_FEATURE_FOO

The scripts/check_kconfig.sh could also just use an environment variable.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>