gen-dynamic-pci.py: provide IOMMU group association
Sometimes you may see an error starting qemu saying that you should
make sure the IOMMU group matches, this typically is the case when you
try to PCI-E passthrough devices but forgot to also passthrough other
related devices in the same IOMMU group.

Provide the IOMMU group in the description then, that way it is very
easy to determine associated devices.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>