gen-dynamic-pci.py: add kconfig symbols for names for devices
We have an association of the kconfig symbols to specific details of a
device, but we don't have the human-readible name, or even the generic
PCI-ID. We can easily grow this and add support for this so that it is
easier for us to do backward inference on ansible variables.

Also do the part to pass these onto extra_vars.yaml so we can use these
as neeeded on ansible roles. We end up with things like these on the
extra_vars.yaml:

pcie_passthrough_devices:
  - { domain: "0x0000", bus: "0x00", slot: "0x1f", function: "0x0",
      pcie_id: "0000:00:1f.0",  iommu_group: "12"
      sdevice: "Tiger Lake-LP LPC Controller",
      pcie_human_name: "Tiger Lake-LP LPC Controller" }
  - { domain: "0x0000", bus: "0x00", slot: "0x1f", function: "0x3",
      pcie_id: "0000:00:1f.3",  iommu_group: "12"
      sdevice: "Tiger Lake-LP Smart Sound Technology Audio Controller",
      pcie_human_name: "Tiger Lake-LP Smart Sound Technology Audio Controller" }
  - { domain: "0x0000", bus: "0x00", slot: "0x1f", function: "0x4",
      pcie_id: "0000:00:1f.4",  iommu_group: "12"
      sdevice: "Tiger Lake-LP SMBus Controller",
      pcie_human_name: "Tiger Lake-LP SMBus Controller" }
  - { domain: "0x0000", bus: "0x00", slot: "0x1f", function: "0x5",
      pcie_id: "0000:00:1f.5",  iommu_group: "12"
      sdevice: "Tiger Lake-LP SPI Controller",
      pcie_human_name: "Tiger Lake-LP SPI Controller" }

The pcie_human_name is enhancements to the names but is subsystem
specific, we only have enhanced support for now for NVMe drives.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>