libvirt: move driver override/unbind work to ansible role
We currently do the libvirt permission managemet for the sysfs
PCI-E device files for driver override / unbind when we create the
extra_vars.yaml file. We did this out of pure convenience as we
already had a loop which iterated over target PCI-E devices, but
there is no need to do that now that we have a proper ansible
variable which keeps track of all the PCI-E devices we wish to
passhtrough under pcie_passthrough_devices, so just use that.

This makes it easier to read / manage and scale this, should we need
to do that. This should also make it easier to manage to do this with
other virtualization technologies.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>