dynamic-pcie: add fix for regular make menuconfig
Jeff Layton reported that just running `make menuconfig` is broken
on Fedora. Oddly enough this happens because the generated file is
only generated when you enable dynamic-kconfig using an environment
variable, this file is only conditionally included in kconfig logic
whent he same variable is defined. The odd thing about this is that
although I knew long ago this issue existed, and it is why the old
HAVE_SUSE_VAGRANT is implemented using a check for -s, an empty file.
However, what makes this odd is I was not seeing the file requirement
to exist *today* on Debian even if an enironment variable is not set
and kconfig does not include it in that context. This may mean the
requirement issue is a library thing.

Fix this by just always the file always, in both worlds:

 * where its needed for dynamic kconfig
 * when its not needed, just touch it

Also just add it to the list of files cleaned up via 'make mrproper'.

Reported-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>