gen_hosts: add selftests support for hosts and nodes file generation
When you don't have a dedicated workflow, ie when the kconfig symbol
CONFIG_WORKFLOWS_DEDICATED_WORKFLOW is not set or "n", then a generic
hosts file for ansible and vagrant/kdevops_nodes.yaml file is used.

When you a dedicated workflow (CONFIG_WORKFLOWS_DEDICATED_WORKFLOW=y)
for a workflow you are expected to write the respective workflow
template file and ansible tasks which let you infer the host / nodes
file for that workflow. This is because a dedicated workflow is intended
to be guided to create as many guests as possible to break down work
into as many guests / target nodes as possible so to break down testsing
from hat would otherwise be parallel a serial workflow to a parallelized
workflow.

Off course the parallelized workflow is completely optional, to
parallelize a workflow is just an option. A perfect example is that
of the Linux kernel selftests. Sure, we can tests selftests all in one
shot, but we can make them faster by splitting target tests by target
selftests as well, this way we parallelize known selftests to run into
each target node.

The are two files we need to generates:

1) the kdevops_hosts_template generation (ansible hosts file)
2) the kdevops_nodes_template generation (vagrant/kdevops_nodes.yaml) which
   can be used by things like libvirt to specify  how we want to create guests

The question is how to easily scale allowing different workflows to
aim at supporting dedicated workflows with the least amount of code
changes. The way to do this for fstests and blktests is by allowing
each workflow to have a dedicated namespace on the configuration file,
and so we use for example all CONFIG_SELFTESTS_SECTION_* entries as
possible target nodes / tests. This let's us simplify things by allowing
the last part of this name to represent the target test name which we
later will use a directory under Linux toolts/testing/selftests/. The fact
that there is a directory that exist for each in toolts/testing/selftests/
is just a convenience for us, we can have our own custom mapping if
that does not suffice as well.

We essentially grep for the enabled selftests and create a lower case
version of the variables to then iterate on the jinja2 template file
over them to create entries for each. This is in essence the target
goal for the hosts file:

{% for test_type in selftests_enabled_test_types %}
{{ kdevops_host_prefix }}-{{ test_type }}
{% if kdevops_baseline_and_dev %}
{{ kdevops_host_prefix }}-{{ test_type }}-dev

And we strive to do this with as little effect to code as possible, just
a set of standard kconfig symbols for the workflow. This is a trick
picked up and learned by fstest and blktests, each dedicated workflow
can pick their own strategy.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>