bootlinux: enable module signing / and zstd compression
All of our kernels have had CONFIG_MODULE_SIG_KEY="" and so although
we enable support for kernel module signing the key is empty and no
modules get signed, and so all of our kernels taint on the load of
any of our modules.

To help avoid issues with future testing let's ensure we just enable
module signing as most folks do today and use the default snake-oil
cert certs/signing_key.pem. This will trigger the generation of the
snake-oil at build time.

Module signing however is done when CONFIG_MODULE_SIG_ALL is enabled,
but this is done make modules_install target time. This works for
all cases except if you are using 9p on the host and mounted the Linux
directory via 9pfs on the client. In such case the certs/signing_key.pem
is owner read only, so 400, we must change it to 755 to ensure the
virtualized guest can access it, so we add a few stub rules for that.

While at it, as we are standardizing with most distros use enable zstd
module compression.

Note that the hunt for the snake-oil is completely optional, in that
if it is not found we ignore it and hope you have done what is right
for your configuration. In the future we may want to just read the
config file and extract the CONFIG_MODULE_SIG_KEY from it and use
that instead.

We only enable the snake-oil and next-20221118 for next-20221118.
Other kernels will have to slowly pick this up, but the hope is that
this will trickle down to all of our other future kernel configs.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>